画像の名前を変更

+++ nextjs-laravel/ecsite/server/database/migrations/2024_10_04_024408_stars.php
@@ -14,9 +14,9 @@ return new class extends Migration
         Schema::create('stars', function (Blueprint $table) {
             $table->id();
             $table->integer('star')->default(0);
-            $table->integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->timestamps();
         });
