+++ nextjs-laravel/ecsite/server/database/migrations/2024_10_04_025911_purchase_historys.php
@@ -12,9 +12,9 @@ return new class extends Migration
     {
         Schema::create('purchase_historys', function ($table) {
             $table->id();
-            $table->integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->timestamps();
         });
